feat: remove tenancy rules
The large number of cases where resources do not fit in the /orgs or
/groups tenancy path hierarchy indicates that this is not a good rule.

Endpoints that depart from this convention will need to define how they
relate to orgs and groups for the purpose of authorization, but that is
not an invariant that can be unconditionally applied to all resources.

Attempting to, and making constant exceptions in reaction to special
cases, is only avoiding the reality that there needs to be a more
generally useful way to express authorization and capability context
outside a strict path hierarchy.