fix: woollypully test failures in dev environments
woollypully tests execute sweater-comb/optic-ci from a relative module
path. This works in development as well as in a packaged OCI image, but
in development, it only works if the entrypoint script is tsc-compiled
first.

`yarn build` is chained with `jest` in the `test` script to ensure that
this entrypoint is compiled before tests run. This introduces minimal
build time overhead as `build` will only compile when output is missing
or out-of-date.